Finish increase view by LEO

+++ src/main/java/enclave/com/service/FilmService.java
@@ -27,6 +27,6 @@ public interface FilmService {
 	Film updateFilm(Film film);
 	
 	List<Film> searchFilmByName(String name);
-	Film setView(Integer id);
+	Film setView(Long id);
 	/* end*/
 }
